refactor: 6054 Update WMGv2 Main (#6257)
